Add backends with uniform interface similar to Vc
This adds backends with wrappers around standard scalar types, so that
the same syntax can be used with them as is used with Vc vector types.
This helps writing generic code for scalar and vector implementations.

The math functions are intentionally ommitted, since they are provided
either by the standard library, or the Vc library. Only functions that
are not provided must be implemented in the backend itself. Also absent
are CondAssign() and MaskedAssign() functions, since now one can use
var(mask) = value; even for the scalar backend. Finally, mask types are
associated with vector types, so for Float_v there is Float_v::Mask, and
so on. Same thing for the vector size, Float_v::Size should give the
value. Thay way the backend namespace remains cleaner.