Prefer LLVM standard C++ library on Darwin
CDash is showing errors like the one below for quite some time for
AppleClang 5.1:

/.../macitois17-clang-geant/USolids/../base/Global.h:228:15: error: no
member named 'unique_ptr' in namespace 'std'

This is because it is using gcc-4.2 standard C++ library, which on
Darwin doesn't have support for C++11 yet. For AppleClang 6.0, this has
been fixed, but I am manually adding this option for Clang builds to fix
this error for the earlier versions of the compiler.