[VecGeom-197] Fixes to CUDA compilation
Details:
CUDA compilation was broken since lima/jira-151 branch was
merged into master, because of inconsistent CUDA treatment of new
setters created for Geant4 integration.

Geometry elements are not supposed to be altered by the GPUs, hence
all setters and other similar functions are now available only on
the host.