Make single precision the default in VecCore
Since Vc doesn't support vectors of 64 bit types to match Int_v and
Real_v in the backend implementations with the same number of elements,
we shall default to single precision during the backend specification
and the creation of conformance tests. Once Vc adds support for vectors
of 64 bit integer types, the default for Real_v can be changed back to
double precision.