[VecGeom-221]  Several fixes to DistanceToin and DistanceToOut
Details:

The addition of convention checker and the use of ShapeTester for the tube showed
several bugs in the tube implementation, mostly on the proper treatment of boundaries
and wrong-side points.

Some of the fixes were hard to implement without new code to identify tracks entering
or leaving the shape.  Some of the code was rearranged and comments added to simplify
future code maintenance.

With these fixes, a simple (solid, non-hollow) tube passes all conventions and shape
tests.  More tests are needed for hollow tubes and tubes with phi-sections.