VECGEOM-35 Allow cascading cuda libraries
In order to allow cascading cuda libraries, we need to have a simple
version of the libraries which contains solely the .o files corresponding
to their content.

Then once the full set of libraries is known, we need to generate a single
intermediary 'link' .o file, generated with nvcc with the -dlink options
that is linked against the .o of the most derived libraries and the 'simple
version' of the dependent libraries.

The final useable library is named lib[targetname]_final.a.  To use
do something like:

  cuda_add_library(cudauserlib ${SRC_USER_CUDA_CPP} OPTIONS ${CUDA_ARCH} )
  cuda_add_library_depend(cudauserlib libvecgeomcuda.a)
  target_link_libraries(cudauserlib ${VECGEOM_LIBRARIES} ${VECGEOM_LIBRARIES_EXTERNAL})
  set(USERKERNELLIB cudauserlib_final)