[VECGEOM-143] Adapting ShapeTester to the VecGeom-USolids policy
* Details:

When VecGeom shapes are used with Geant4 environment, negative distances
or safeties cause problems in Geant4.  [Scalar]ShapeImplementationHelper
uses macro VECGEOM_REPLACE_USOLIDS to decide whether the negative values
from VecGeom conventions for distances and safeties are converted into
zeros, as mandated in USolids conventions.

ShapeTester (and CoventionChecker) are adjusted to test the conventions
for both situations (VECGEOM_REPLACE_USOLIDS=ON or OFF) as needed.