VECGEOM-35 Quick fix for stable address of gCompactPlacedVolBuffer on device.
The underlying problem is the fact that GeantV currently (ab)use the CUDA single compilation unit mode which prevent
the use of extern variable.  GeantV pretends to use the single compilation unit mode but still relies on libvecgeomcuda.
This needs to be fixed but changing the compilation step as proven to be 'fragile', so instead we replace the
global variable by a function static (shared) variable (which anyway is usually a better choice).