VECGEOM-35 Allow cascading cuda shared libraries
Add support for cascading shared library in the same way as for static
library.

Instead of libname.a, a static library named libname_static.a is created.
Instead of libname_final.a a shared library named libname.so is created.

To use, add something like:

  cuda_add_library(cudauserlib ${SRC_USER_CUDA_CPP} SHARED OPTIONS ${CUDA_ARCH} )
  cuda_add_library_depend(cudauserlib vecgeomcuda_static libvecgeomcuda_static.a)

  target_link_libraries(cudauserlib ${VECGEOM_LIBRARIES} ${VECGEOM_LIBRARIES_EXTERNAL})
  set(USERKERNELLIB cudauserlib)

Note: the repeat in cuda_add_library_depend below is due the author of cuda_add_library_depend
not knowing how to automatically go from the target to the real file in add_custom_command.