Use standard exceptions to exit from unimplemented functions
The usage of pragma message in various places creates warnings on CDash,
and does not fail when an unimplemented function is called, which might
lead to bugs that are very hard to debug. In order to clean up build
warnings, this commit exchanges the usage of pragma message with code
that will throw an exception during runtime and properly fail when the
unimplemented functions are called.