[VECCORE-13] Update VecGeom's CMakeLists.txt for VecCore changes
The SIMD=ON/OFF option of VecCore is not needed anymore. Since VecCore
is now always needed by VecGeom, we remove the ENABLE_VECCORE option to
enable building VecCore and just add the subdirectory by default. The
include_directories() commands are still needed as they are not
propagated to parent directories by CMake.