Move VecCore default compile options to CompilerSetup.cmake
This moves the default compile options for release, debug, and
release with debug info builds to the common CompilerSetup.cmake
module. It also drops defining the DEBUG macro, as that currently
enables broken code in USolids. Compiler-specific compile options
are still possible, by just overwriting the default settings in
the CMake module for the particular compiler.