Fix bug in HypeImplementation.h
Using typedefs to create a local Float_t type is a bug and must
be banned from being used in VecGeom. The type Float_t is defined
by ROOT at global scope and conflicts with this sort of construct,
which create inumerous places where scalar and vector types are
being mixed in the code. This is an issue that must be addressed
urgently in the whole of VecGeom by eliminating all of such constructs
with similar changes to this commit. Below is the error that led me
to identify this problem:

VecGeom/volumes/kernel/HypeImplementation.h(346):
error: no instance of overloaded function "vecgeom::cxx::MaskedAssign"
matches the argument list
            argument types are: (UME::SIMD::SIMDVecMask<4U>,
	    UME::SIMD::SIMDVec_f<double, 4U>, Float_t *)
   MaskedAssign(pz < 0., -vz, &vz);
   ^
This means that the compiler could not identify Float_t* as being
a pointer to typename Backend::precision_v (a.k.a.
UME::SIMD::SIMDVec_f<double, 4U>), in this context.