Fix bug in ShapeImplementationHelper class
Cannot use MaskedAssign here, as compiler cannot deduce proper vector
type from a pointer to the scalar type without a cast. We might need a
MaskStore function in VecCore's interface for this sort of operation.
Note also that daughterId is a scalar here, rather than the proper type
Backend::int_v. The compiler cannot guess such conversions, so a call to
the constructor to convert is absolutely required in this context.
Worse, if a conversion from mask to bool is defined for the mask type,
then that will be used to call the scalar version of MaskedAssign, which
will only ever assign to the first element starting from the pointer,
while leaving other elements unchanged, which is a very serious bug.

VecGeom/volumes/ShapeImplementationHelper.h(480):
error: no instance of overloaded function "vecgeom::cxx::MaskedAssign"
matches the argument list
            argument types are: (vecgeom::cxx::UMESIMDMask, int, int *)
    MaskedAssign(valid, daughterId, nextDaughterIdList + i);
    ^