[VECGEOM-375] Improving robustness of trapezoid in construction of side planes
Details:

A recurring trapezoid problem was caused by trap parameters producing non-planar faces.
An easy demonstrator is the TestVECGEOM-375() function in TestTrap unit test.  Points
provided by GetPointOnSurface() were frequently reported as not being on surface.  An
easy fix, although very fragile, was to relax the tolerance for the trapezoid surface.

This commit fixes the issue in a robust way, with a new method fromPlanesToCorners(),
which provides corner points based on the actual intersection of constructed side planes,
and using these co-planar corners in GetPointOnSurface(), Extent() and Capacity().

Warnings are printed when non-planar corners are associated to user-provided parameters
(via fromParametersToCorners() function), but if residuals are not too large (currently
1000*kTolerance), the plane faces are used anyway.