[VECGEOM-315] Generic factory mechanism for unplaced volumes
Implementation of a central entry point to create instances of
UnplacedVolumes:

auto tube = GeoManager::MakeInstance<UnplacedTube>(rmin, rmax, ...)
auto box  = GeoManager::MakeInstance<UnplacedBox>(x,y,z);

The benefit of this interface is that the user might get a specialized
instance of a subtype (like a FullTube instead of a GenericTube).

The new interface should be from now on the generic API to construct unplaced
volumes. We could eventually disallow creation via standard constructors for
the benefit of
  - automatic ownership handling
  - optimized memory placement/allocation etc

This commit further provides

- special implementation for MakeInstance of tubes to be served
  as template for other shapes which can follow
- a unit test
- use of MakeInstance<> already in a couple of places like Root-VecGeom conversion