[VECGEOM-476] Prevent MakeMinusTolerantSquare to get negative
The MakeMinusToleratnSquare function is supposed to calculate
(x-tol)*(x-tol) which should never be smaller than 0.

But since we neglect the second order tol*tol term in order to be fast,
this condition might exceptionally be violated.

This commit works around this problem by
taking Abs( ) of the final result. This fixes the failing
unit test reported in [VECGEOM-476].