Offer possibility to convert units to MM when importing from TGeo
Extend the RootGeoManager with the possibility to perform
a scale conversion of lengths during converion TGeo -> VecGeom.

A VecGeom geometry can than be directly compatible with G4 requirements.

In order to activate this conversion, a user should do

vecgeom::RootGeoManager::Instance().EnableG4Units();

(but only if the original units are in cm).

By default, this is not activated, corresponding to

vecgeom::RootGeoManager::Instance().EnableTGeoUnits();