Introducing Intel Embree
This introduces **optionally** the possibility to use
Intel Embree inside VecGeom. It serves as the initial seed for
further extensions and developments.

The commit introduces an EmbreeNavigator which might be used in
tessellated solids where both the initialisation time and the runtime improve
(for large N).

This commit provides the code used for the first investigation:

https://indico.cern.ch/event/732271/contributions/3019459/attachments/1657103/2653198/Embree_Investigation_SWenzel.pdf