[VECGEOM-486]: Less template specialization for Polyhedron
The polyhedron has too many template (volume) specializations which
was a heavy burden for compilers.
This commits disables the specializations on placement as a quick fix
of [VECGEOM-486].

Nevertheless, I am in favour of (re-)allowing few important placement
specializations (no translation, no rotation) in the future which should
still be managable.