[VECGEOM-494] Introduce enhanced volume specialization capabilities
This commit introduces a template class (proxy) UnplacedImplAs<A,B>
which implements an unplaced volume of type A behaving/implemented as
B.

The commit contains several accompanying changes/modifications to make
the complete specialization workflow, including placement of such objects
work correctly.

The mechanism is demonstrated for the sphere (in SphereBenchmark) which
can now generate an orb-behaving object.

As a side note, the commit starts migrating away (the implementation of)
ROOT+Geant4 conversion functions from the placed volumes to the unplaced volumes
since this makes our life much easier.