Add FAST_MATH option.  Default to off.
It does not work with SSE (at the very least comparision between a number and infinity
in vector mode give the wrong result).   It can now be enabled on demand to resolve
this issues or test for performance.

Note detecting for SSE from CMake is not straight-forward in case where VECGEOM_VECTOR
is set to native (i.e. the default).