Remove internal build of XercesC
As GDML is an optional component and Xerces is a common and easy
to build external, remove ability to build it as an external.

Make find of external Xerces required. Link vgdml to imported target
supplied by CMake's FindXerces module.

Remove un-needed GLOBing of sources as only have a few sources, so clearer
to list explicitly.

At present, there is no export of the vgdml target, as neither are the
core VecGeom targets.