Use GNUInstallDirs to define install locations
Replace custom install location variables with standard and FHS
compilant set provided by CMake's builtin GNUInstallDirs module.

Provide a basic set of CMake settings to check for absolute install
locations and to reduce verbosity of install messages when files
have not changed.

Replace all occurences (except Jenkins and RecordCmdLine.cmake) of
CMAKE_{SOURCE,BINARY}_DIR variables with PROJECT_{SOURCE,BINARY}_DIR.
The PROJECT versions are more robust under the use case that VecGeom
is imported/built as a submodule of a downstream project. Jenkins
scripts are left as-is as they are project specific. RecordCmdLine.cmake
is unchanged as it must run before project() is called and thus the
PROJECT variables do not exist.

RecordCmdLine is called before Project