Added index-based navigation as alternative to the standard path-based NavigationState.
Compilation using DUSE_NAVINDEX=ON. A navigation index table is built in this case by GeoManager::CloseGeometry. The table stores the navigation index for all physical volumes, including the relationship parent-child and the cached value of the global transformation to the node, down to a geometry depth specified via GeoManager::SetTransformationCacheDepth. The default is 0 meaning caching down to full depth.