Force paths to refind builtin VecCore
Reported by @gcosmo when using VecGeom in a Geant4 build - when finding
VecGeom via VecGeom_DIR, builtin VecCore is not found.

As we know when we use and where a builtin VecCore will be installed,
export this knowledge to VecGeomConfig.cmake. Force find of VecCore
via find_package's override mechanism so that only the installed builtin
path should be considered.

Note that this only covers this one use case and no further downstream
"builtin" dependencies. Have to rely on clients setting <package>_DIR
or CMAKE_PREFIX_PATH accordingly either directly or via environment.