Always check and download test data
The DATA_DOWNLOAD option for testing is effectively unused because

- Testing always(*) requires the data, downloaded or locally copied
  by the client.
- Download is only triggered if the data is not present.
- Tests will fail if the data isn't present

(*) only if the full suite of tests is enabled. It may be possible
to restrict check/download for subsets of tests.

Remove the DATA_DOWNLOAD option, always checking for the data files
and downloading if not present or there is a hash mismatch. Rewrite
download/hashing to use CMake commands only.