Expose GDML user and volume auxiliary data
The GDML schema provides for arbitrary data to be attached to <volume>
elements and the <userinfo> element through 0-N <auxiliary> tags. This
tag can itself can hold 0-N child <auxiliary> tags, thereby modeling a
tree structure. This structure, and the values of the attributes of
the tag attributes are left up to the client.

Process and expose any volume and userinfo data in an input GDML file
through vgdml's Middleware interface. Provide a class modelling the
<auxillary> tag/tree structure, with interfaces in Middleware to
expose:

- A map between VecGeom logical volume ids (int) holding auxiliary data
  and the list of the held auxiliary tags
- A list of auxiliary tags in the <userinfo> element, if any.

Implement basic tests of Auxiliary class and correct extraction of data
from GDML by Middleware