Do not fallback to builtin VecCore if external not found
Reported in VecGeom-583. We should not fallback to the builtin VecGeom
option if use of external was requested but could not be found.

Remove fallback to builtin VecGeom when external package requested and
not found.

Support switch between builtin/external in current build directory to
ensure consistent builds.

TODO: Use of FetchContent in CMake >= 3.11 is preferred, but VecCore
does not support this well yet as it is not designed to be used as
a subproject.