Explicitly use system VecCore in Jenkins builds
Jenkins builds use an LCG view for build dependencies, which includes
an install of VecCore. With the default being to use a system install
on VecCore, the removal of the fallback to builtin does not change
the behaviour here.

Add BUILTIN_VECCORE=OFF to default build arguments to make this use
case explicit.