[CMake] Step back to single, static CUDA library
Provide a baseline build using canonical static non-device-linked
library for VecGeom CUDA functionality. Using the library, modulo
circular dependencies between vecgeom<->vecgeomcuda, then works
simply:

- Client is shared/static library with private CUDA use: link and device link
- Client is executable: link and device link
- Client is static library with public CUDA interface: link

Use in a shared library with public CUDA (i.e. device) interface is not
supported. This baseline is to provide a starting point for how, or if,
this can be supported.