Provide static and shared variants of libvecgeomcuda
Use native CMake CUDA support to build static and shared variants
of the vecgeomcuda library. This reproduces the behaviour of the
hacked FindCUDA module removed in the migration to native CMake
support:

- The static variant is a "standard" CUDA library that consumers
  can link/device link to.
- The shared variant includes a device link object so that non-CUDA
  consumers can link to it without requiring an otherwise unneeded
  device link.
- CUDA consumers can link to the shared variant provided they
  device-link explicitly to the static vecgeomcuda.

Do not link the VGDML library to either CUDA variant of vecgeomcuda,
allowing clients to use the desired static/shared variant.

Update tests to use only static variant of vecgeomcuda when available.