Drop linking libraries from custom ClangTidy check
This is a shared object that gets loaded by clang-tidy,
so it doesn't need to be linked against clang libraries.
In clang 10 and above libraries may be merged into a single
libclang.so as well, so in this case the current settings
failed to build the plugin as well, because it cannot find
the split libraries.