Fix size calculation of navigation index table
The previous commit 5c505fb4 "Ensure that transformation data in nav table is stored aligned" did not update the navigation index table size correctly. This lead to out of bounds access, because the allocated buffer for the table was not large enough. Commit b99894e0 fixed an assertion triggerd by this bug, but did not fix the issue for all code paths.