[cuda] Expose function to change heap limit
As of now, all "dynamically sized" shapes such as polyhedra, polycones,
and extruded solids allocate memory in kernels. This is problematic as
CUDA only has a very limited amount of scratch space for malloc / new.
Until the "transfer code" for these types of shapes is fixed, add the
function CudaDeviceSetHeapLimit similar to CudaDeviceSetStackLimit to
let VecGeom clients work around this problem.