Simplification of daughter crossing check.
Simplify the current logic which uses the mother volume convexity to prioritize daughter checks. This approach is not always justified because it may increase the search range for daughter candidates. The current approach is conservative and uses the distance to mother first to limit the search distance for daughters, only if there are daughters. The check for daughters at this point spares extra checks in the derived navigators, which in the case of BVH was even generating a crash.