Minimise use of ROOT/Geant4 in main VecGeom library
* CMake genexes should be quoted to support older CMake

* Remove unneeded extra link

* Fixup config/build errors for GCC/CUDA

* Create VecGeomTest library for all shared testing code

Supercedes previous testing libraries to provide single point of
contact for all VecGeom tests. Simplifies linking for now, pending
longer term review of required testing functionality.

* Minimize ROOT/Geant4 link footprint in VecGeom library

* VGM is only needed by G4GeoManager as private dependency

* Make GDML tests a subdirectory of main tests

- As one tests relies on ROOT support, want to manage it here
  because we only support ROOT in tests.
- Cleaner organization, find all tests under one tree.

* Initial hacky move of RootGeoManager and Root shapes into tests

As with G4GeoManager, these classes are only needed for testing so we
move it to tests. Compile, temporarily, into the testing utilities
library for easiest move. Update tests to include right header.

TODO: GDML testing relies on this, so those tests will not build
at the moment.

* Isolate ConvertTo{Geant4,Root} functions as far as possible

Pure semi-tidy to reduce footprint of VECGEOM_{GEANT4,ROOT} symbols
to assist in later removal of these interfaces:

- Don't include Geant4/ROOT headers where not needed.
- If ConvertTo function can be inlined in class body without
  requiring inclusion of Geant4/ROOT headers, do so.

* Move G4GeoManager out of VecGeom interface

This class is only used for global benchmarks and should not be
part of the main VecGeom interface. Should the functionality be required
upstream, it should be placed into a separate library to avoid coupling
VecGeom itself to Geant4. This would create a circular dependency when Geant4
has VecGeom solid use enabled.

* Remove unused ResultComparator includes

* Transformation3D::ConvertToTGeoMatrix can be static

Only requires data obtainable from public interface. Update client
code to use the static method.

* Fully remove ROOT dependence in NavState{Index,Path}

Migrate remaining ROOT conversion function to free function only
used in testing code. Update tests using this function with the
relevant header and call.

* Remove NavState{Index,Path}::GetCurrentNode

Per the comments, this function was purely for GeantV and should
be removed.

* NavState{Index,Path} ROOT conversion/access can be static

- Remove copy assignment from TGeoBranchArray as this is not used
- Make conversion/access ROOT functions static, as they can be
  implemented using public methods
- Update tests to use static member functions

* Remove ROOT dependence of FlatVoxelHashMap

Persistency functionality never used in testing other than to
write/readback, so no currently useful production reason for this.
Clients requiring persistency can easily add their own binding if
required.

* Minimize ROOT use in FlatVoxelManager

Remove private functions that are never called anywhere. Migrate
remaining static functions to free functions in anonymous namespace
to hide as much as possible.

TODO: Clarify the use, or not, of EXTREMELOOKUP which is the only
use of the dump to TFile calls.

* Make all Volume use of VECGEOM_{ROOT,GEANT4} consistent

* Remove ROOT usage which is no longer relevant

* Code never uses ROOT, so remove VECGEOM_ROOT

* Remove unused, obsolete header