PlaneShell: Remove broken optimization for Trapezoid
PlaneShell's SafetyToIn and SafetyToOut contained what looks like an
optimization for the case N == 4. However, the condition / branches
were swapped so that the code would use the general loop for N == 4
and assume exactly four planes for N != 4. This wasn't a problem in
practice because VecGeom only ever uses PlaneShell with N == 4, for
the Trapezoid.

Performance measurements show that the re-ordering of Max operations
yields a very minor improvement of at most 1.5% - 2%. As it wasn't
active up to now, just remove the code path. As such, no performance
difference expected nor measured.